pyproject.toml
[project]
name = "mlir-rl-artifact"
version = "0.1.0"
description = "A deep reinforcement learning system for loop nest optimization in MLIR"
authors = [
    {name = "Mohammed Tirichine",email = "km_tirichine@esi.dz"},
    {name = "Nassim Ameur",email = "kn_ameur@esi.dz"},
    {name = "Iheb Nassim Aouadj",email = "nassimiheb.aouadj@gmail.com"},
    {name = "Nazim Bendib",email = "jn_bendib@esi.dz"},
    {name = "Bouchama Djad",email = "bouchamadjad@gmail.com"},
    {name = "Rafik Bouloudene",email = "rafikobouloudene@gmail.com"},
    {name = "Riyadh Baghdadi",email = "baghdadi@nyu.edu"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy (>=1.19.5,<=1.26)",
    "tqdm (>=4.67.1,<5.0.0)",
    "typeguard (>=4.4.4,<5.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "monkeytype (>=23.3.0,<24.0.0)",
    "pyyaml (>=5.3.1,<=6.0.1)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "markdown-pdf (>=1.10,<2.0)",
    "torch (>=2.8.0,<3.0.0)",
    "torchvision (>=0.23.0,<0.24.0)"
]


[project.scripts]
train = 'train:main'
evaluate = 'evaluate:main'


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "explicit"

[tool.poetry.dependencies]
torch = {source = "pytorch-cpu"}
torchvision = {source = "pytorch-cpu"}
